update readme and remove redundant plugins
