+++ demo/index.html
@@ -26,13 +26,10 @@
     <script src="/demo/lib/matter-tools.demo.js"></script>
 
     <!-- Plugins -->
-    <script src="../examples/attractorsPlugin.js"></script>
-    <script src="../examples/gravityPlugin.js"></script>
-    <script src="../examples/wrapPlugin.js"></script>
+    <script src="/demo/lib/matter-wrap.js"></script>
     
     <!-- Examples -->
     <script src="../examples/airFriction.js"></script>
-    <script src="../examples/attractors.js"></script>
     <script src="../examples/avalanche.js"></script>
     <script src="../examples/ballPool.js"></script>
     <script src="../examples/bridge.js"></script>
